backend/requirements.txt
fastapi
pydantic
python-multipart
openai
streamlit
numpy
pdfplumber
bcrypt
passlib
python-jose
motor
